requirements.txt
jupyter
matplotlib
pandas
scikit-learn
langchain_openai
langchain_core
python-dotenv